Fix multiple engine build issues:
  * Fix GPG key imports for MySQL failing 50% of the time
  * Pin libpq-dev to use pg11 (in the current version of the Postgres image it pulls in libpq-dev from pg12 and breaks pg_config and all extension builds)
  * Bump postgres to 11.7

Also add a Makefile to build the engine using Buildkit (currently a thin wrapper but also allows targeting just the toolchain stage in order to debug extension builds)

Bump Splitgraph submodule: still not sure what the best practice for this is, do we build the engine in CI, push it to Docker Hub and then test it using Splitgraph integration tests? Do we merge the engine and Splitgraph code into a single repo?